+++ src/main/java/com/example/model/Question.java
@@ -12,7 +12,7 @@ import java.util.Date;
 @Data
 public class Question {
 
-        private Integer id;
+        private Long id;
 
         private String title;
 
@@ -21,11 +21,12 @@ public class Question {
 
         private Long gmtModified;
 
-        private Integer creator;
+        private Long creator;
 
         private Integer commentCount;
 
         private Integer viewCount;
+
 
         private Integer likeCount;
         private String tag;
